Further review corrections.
