+++ src/DirectWebServices/Sectors/SectorsBreakdownOptions.ts
@@ -22,38 +22,6 @@
 export namespace SectorsBreakdown {
 
     /* *
-    *
-    *  Interfaces
-    *
-    * */
-
-    export interface InvestmentResponse {
-        identifiers: {
-            performanceId: string;
-        };
-        metadata: InvestmentMetadata;
-    }
-
-    export interface InvestmentMetadata {
-        requestId: string;
-        time: string;
-        messages?: Array<Message>;
-    }
-
-    export interface Message {
-        type: string;
-        investments?: Array<InvestmentMessage>;
-    }
-
-    export interface InvestmentMessage {
-        id?: string;
-        idType?: string;
-        status?: string;
-        datapointId?: Array<string>;
-        errorCode?: string;
-    }
-
-    /* *
      *
      *  Constants
      *
